refactor(input): Barrel file renaming
