+++ projects/igniteui-angular/src/lib/date-picker/date-picker.component.spec.ts
@@ -7,7 +7,7 @@ import { IgxDatePickerComponent, IgxDatePickerModule } from './date-picker.compo
 import { IgxLabelDirective } from '../directives/label/label.directive';
 import { IgxInputDirective, IgxInputState } from '../directives/input/input.directive';
 import { UIInteractions, wait } from '../test-utils/ui-interactions.spec';
-import { IgxInputGroupModule, IgxInputGroupComponent } from '../input-group';
+import { IgxInputGroupModule, IgxInputGroupComponent } from '../input-group/public_api';
 import { IgxTextSelectionModule } from '../directives/text-selection/text-selection.directive';
 import { configureTestSuite } from '../test-utils/configure-suite';
 import { IgxButtonModule } from '../directives/button/button.directive';
