+++ projects/igniteui-angular/src/lib/date-range-picker/date-range-picker-inputs.common.ts
@@ -1,7 +1,7 @@
 import { Component, ContentChild, Pipe, PipeTransform, Output, EventEmitter, HostListener, Directive } from '@angular/core';
 import { formatDate } from '@angular/common';
 import { NgControl } from '@angular/forms';
-import { IgxInputDirective, IgxInputState } from '../input-group/index';
+import { IgxInputDirective, IgxInputState } from '../input-group/public_api';
 import { IgxInputGroupComponent } from '../input-group/input-group.component';
 import { IgxInputGroupBase } from '../input-group/input-group.common';
 import { IgxDateTimeEditorDirective } from '../directives/date-time-editor/public_api';
