+++ projects/igniteui-angular/src/lib/date-range-picker/date-range-picker.component.ts
@@ -41,7 +41,7 @@ import { IToggleView } from '../core/navigation';
 import { CancelableBrowserEventArgs, IBaseEventArgs, KEYS } from '../core/utils';
 import { DatePickerUtil } from '../date-picker/date-picker.utils';
 import { IgxToggleDirective } from '../directives/toggle/toggle.directive';
-import { IgxInputDirective, IgxInputGroupComponent, IgxInputState, IgxLabelDirective } from '../input-group/index';
+import { IgxInputDirective, IgxInputGroupComponent, IgxInputState, IgxLabelDirective } from '../input-group/public_api';
 import { AutoPositionStrategy, OverlaySettings, PositionSettings } from '../services/index';
 import {
     DateRange,
