+++ projects/igniteui-angular/src/lib/input-group/public_api.ts
